refactor(core): split paper queries into submodules
Cut storage/queries/paper.rs into read/write/roots/pdf/fts/trash.
Public API unchanged via mod.rs re-exports; bodies moved verbatim.
